# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WorkLink",
        "description": "Documentación de la API REST",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refrescar token",
                "description": "Genera un nuevo token JWT",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "Token refrescado exitosamente"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/contract-requests": {
            "get": {
                "tags": [
                    "Contract Requests"
                ],
                "summary": "Listar solicitudes de contratación",
                "description": "Retorna las solicitudes de contratación según el usuario autenticado y su rol.",
                "operationId": "listContractRequests",
                "responses": {
                    "200": {
                        "description": "Solicitudes obtenidas exitosamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contract Requests"
                ],
                "summary": "Crear solicitud de contratación",
                "description": "Crea una solicitud para contratar un servicio. Clientes y empresas crean solicitudes; admin puede indicar client_id.",
                "operationId": "createContractRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id",
                                    "description"
                                ],
                                "properties": {
                                    "client_id": {
                                        "description": "Solo requerido si el usuario autenticado es admin",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Necesito una API REST para mi plataforma."
                                    },
                                    "budget": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud creada exitosamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contract-requests/{id}": {
            "get": {
                "tags": [
                    "Contract Requests"
                ],
                "summary": "Obtener solicitud por ID",
                "description": "Retorna los detalles de una solicitud de contratación.",
                "operationId": "showContractRequest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud obtenida exitosamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Contract Requests"
                ],
                "summary": "Actualizar solicitud de contratación",
                "description": "Actualiza la descripción, presupuesto o estado de una solicitud.",
                "operationId": "updateContractRequest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada."
                                    },
                                    "budget": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 700
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "accepted",
                                            "rejected",
                                            "canceled"
                                        ],
                                        "example": "accepted"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Solicitud actualizada correctamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contract Requests"
                ],
                "summary": "Eliminar solicitud de contratación",
                "description": "Elimina una solicitud de contratación.",
                "operationId": "deleteContractRequest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud eliminada correctamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}